Remove eye_occlusion_displace multiplication
Hi, the unity shader graph expects Eye Occlusion Displace values from interval (-1, 1). In Blender, the value is set to 0.5 by default and when exporting, it is multiplied by 100 resulting in a value of 50. This value than has to be re-initialized each time a character is imported into unity. I did not figure out the reason for the division and multiplication, but the removal solves the problem for us. Thanks for looking into it.